capability api refactoring
